# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API do Observatório Turístico de Cumuruxatiba
  version: 0.7.0
  license:
    name: Licença a definir pelo Município de Prado
    identifier: LicenseRef-Proprietary
  description: >
    Contrato da fundação, estadias, questionários e métricas públicas. A
    extensão x-cumuru-feature informa a funcionalidade proprietária de cada
    operação — platform, core, questionnaire, analytics ou self-service — e o
    valor deferred marca a operação ainda não implementada. Tudo roda em modo
    PROTOTYPE_ONLY e somente com fixtures fictícias e generalizadas.
servers:
  - url: https://turismo.prado.ba.gov.br/api/v1
...
security:
  - oidc: []
paths:
...
  /platform/readiness:
    get:
      summary: Consultar readiness do PostgreSQL
      tags: [platform]
      security: []
      operationId: getPlatformReadiness
      x-cumuru-feature: platform
      responses:
        "200":
          description: Dependências necessárias estão prontas.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
  /platform/build:
    get:
      summary: Consultar metadados técnicos do build
      tags: [platform]
      operationId: getPlatformBuild
      x-cumuru-feature: platform
      security:
        - oidc: ["platform:read"]
      responses:
        "200":
          description: Metadados técnicos permitidos do build.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildInfo"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
    session:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token opaco emitido por POST /auth/login. O servidor guarda somente o
        SHA-256 do token e o principal derivado tem os mesmos escopos de um
        principal OIDC.
...
  headers:
    RequestId:
      description: Identificador sanitizado de correlação da requisição.
      schema:
        type: string
    NoStore:
      description: Impede armazenamento de respostas operacionais.
      schema:
        type: string
        const: no-store
...
  responses:
...
    Problem:
      description: Erro no formato Problem Details.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalServerProblem:
      description: Erro interno sanitizado no formato Problem Details.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    ReadinessStatus:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          const: ready
    BuildInfo:
      type: object
      required: [version, revision, built_at]
      additionalProperties: false
      properties:
        version:
          type: string
        revision:
          type: string
        built_at:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        invalid_params:
          type: array
          items:
            $ref: "#/components/schemas/InvalidParam"
    InvalidParam:
      type: object
      required: [name, reason]
      additionalProperties: false
      properties:
        name:
          type: string
          description: JSON Pointer ou nome do campo inválido, sem o valor recebido.
        reason:
          type: string
          pattern: "^[a-z0-9-]+$"
...